Move TypeParameterUtils back into internal.amqp
